move new components to core addon and comment out ICon usage
