--- ui/app/components/toolbar-actions.js
@@ -1,3 +1,0 @@
-import OuterHTML from './outer-html';
-
-export default OuterHTML.extend({});
