--- ui/app/components/toolbar.js
@@ -1,6 +1,0 @@
-import OuterHTML from './outer-html';
-
-export default OuterHTML.extend({
-  classNames: ['toolbar'],
-  tagName: 'nav',
-});
