--- ui/app/templates/components/toolbar.hbs
@@ -1,3 +1,0 @@
-<div class="toolbar-scroller">
-  {{yield}}
-</div>
