+++ ui/lib/core/addon/components/toolbar-actions.js
@@ -1,0 +1,6 @@
+import Component from '@ember/component';
+import layout from '../templates/components/toolbar-actions';
+
+export default Component.extend({
+  layout,
+});
