+++ ui/lib/core/addon/components/toolbar.js
@@ -1,0 +1,7 @@
+import Component from '@ember/component';
+import layout from '../templates/components/toolbar';
+
+export default Component.extend({
+  tagName: '',
+  layout,
+});
