+++ ui/lib/core/addon/templates/components/toolbar-link.hbs
@@ -14,5 +14,7 @@
   data-test-secret-backend-configure={{data-test-secret-backend-configure}}
 >
   {{yield}}
+  {{!--
   <ICon @glyph={{glyph}} @size=12 />
+  --}}
 </LinkTo>
