+++ ui/lib/core/addon/templates/components/toolbar.hbs
@@ -1,0 +1,5 @@
+<nav class="toolbar">
+  <div class="toolbar-scroller">
+    {{yield}}
+  </div>
+</nav>
