+++ ui/lib/core/app/components/toolbar-link.js
@@ -1,0 +1,1 @@
+export { default } from 'core/components/toolbar-link';
