+++ ui/lib/replication/addon/templates/mode/secondaries/config-edit.hbs
@@ -33,7 +33,9 @@
       data-test-delete-mount-config=true
     }}
       Delete
+      {{!--
       <ICon @glyph="chevron-right" @size=12 />
+      --}}
     {{/confirm-action}}
   </div>
 </form>
